Implement better viewer tracking that does not require user info like IP or UA
When the user requests a index.m3u8 that request responds with a dummy
m3u8 file that only contains a link to a new m3u8 that has the
hls_ctx parameter appended to it. this gives us very accurate viewer
tracking but without needing any sort of fingerprinting measure.